Add files via upload
This is a small project which takes user's name and age as input. tells whether they can participate in voting or not, if yes asks to vote and gives the final voting results. if they are not eligible to vote, just gives the final output. it takes the votes randomly from random class.